# feat: Add comprehensive performance benchmark suite